requirements.txt
pycryptodome==3.9.0
base58==1.0.0
bitcoin==1.1.42
eth-abi==1.3.0
ethereum==2.3.2
ecdsa>=0.10
jsonrpcclient==2.0.1
jsonrpc-requests==0.4.0
mnemonic==0.18
requests==2.19.1
tronapi==3.0.5
two1==3.10.9
web3==4.8.2
yarl==0.18.0
colander==1.7.0
pytz==2019.1
django==2.2.5
Celery==4.3.0
django-rest-framework==0.1.0
django-otp==0.7.0
django-cors-headers==3.1.0
django-rest-auth==0.9.5
git+https://github.com/SergeiVasilenko/pycoin@master#pycoin=0.90a
